Now put_integer_64_constant takes an INTEGER_64 value.
git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@28665 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/Eiffel/eiffel/com_il_generation/Core/icore.cs
@@ -392,7 +392,7 @@ public interface ICore
 	void PutInteger8Constant( Int32 i  );
 	void PutInteger16Constant( Int32 i  );
 	void PutInteger32Constant( Int32 i  );
-	void PutInteger64Constant( Int32 i  );
+	void PutInteger64Constant( Int64 i  );
 
 	// Put `d' on IL stack.
 	void PutDoubleConstant( double d );
